% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cage_test_data}
\alias{cage_test_data}
\title{CAGE test data}
\format{
A data frame with three variables: \code{gene_id} Gene ID,
\code{symbol38} Enhancer ID, \code{wilcoxtest_cage} result of the test ).
}
\usage{
cage_test_data
}
\description{
Result of the Wilcoxon tests on the CAGE data of the enhancer gene pairs.
}
\keyword{datasets}
