% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_features.R
\name{compute_features}
\alias{compute_features}
\title{Compute features}
\usage{
compute_features(file, metaData, cores)
}
\arguments{
\item{file}{Path to a file with either genes and enhancer pairs or with only
genes, separated by tabs. The genes need to be given as gencode IDs and
the enhancers as ENCODE CCREs IDs.}

\item{metaData}{Dataframe indicating the paths to the ChIP-seq experiments.
More information on the format here \code{crupR::normali}}

\item{cores}{Number of cores to compute the CRUP score features}
}
\value{
A table containting the following computed features :
\itemize{
\item Distance between gene and enhancer
\item CRUP enhancer score for enhancer region, promoter region and the region between the enhancer and the promoter
\item CRUP promoter score for enhancer region, promoter region and the region between the enhancer and the promoter
\item CAGE wilcoxon test score
\item DNAase-seq wilcoxon test score
\item CRUP correlation scores
}
}
\description{
Computes the features needed for the classification step.
}
\examples{
files <- c(
"/project/CRUP_scores/CRUP_scores/ENCODE/Single_ended/Tissues/body_pancreas/H3K4me1/H3K4me1.bam",
"/project/CRUP_scores/CRUP_scores/ENCODE/Single_ended/Tissues/body_pancreas/H3K4me3/H3K4me3.bam",
"/project/CRUP_scores/CRUP_scores/ENCODE/Single_ended/Tissues/body_pancreas/H3K27ac/H3K27ac.bam" )

inputs <- "/project/CRUP_scores/CRUP_scores/ENCODE/Single_ended/Tissues/body_pancreas/Controls/input.bam"

metaData <- data.frame(HM = c("H3K4me1","H3K4me3","H3K27ac"),
                      condition = c(1,1,1), replicate = c(1,1,1),
                      bamFile = files, inputFile = rep(inputs,3))
file <- "/project/CRUP_scores/EPI/example1.txt"
compute_features(file, metaData, cores)


}
