% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crup_cor}
\alias{crup_cor}
\title{CRUP correlations}
\format{
A data frame with five variables: \code{year}, \code{sex},
\code{name}, \code{n} and \code{prop} (\code{n} divided by total number
of applicants in that year, which means proportions are of people of
that sex with that name born in that year).
}
\usage{
crup_cor
}
\description{
Correlation between the CRUP scores of enhancer and genes  in the pairs
}
\keyword{datasets}
