% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crupexp_test_data}
\alias{crupexp_test_data}
\title{CRUP expression test}
\format{
A data frame with three variables: \code{gene_id} Gene ID,
\code{symbol38} Enhancer ID, \code{wilcoxtest_crup_exp} result of the test ).
}
\usage{
crupexp_test_data
}
\description{
Results of the Wilcoxon test on the CRUP scores and RNA expression of the pairs
}
\keyword{datasets}
