% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dhsdhs_test_data}
\alias{dhsdhs_test_data}
\title{DNase-DNase test}
\format{
A data frame with three variables: \code{gene_id} Gene ID,
\code{symbol38} Enhancer ID, \code{wilcoxtest_dhs_dhs} result of the test ).
}
\usage{
dhsdhs_test_data
}
\description{
Results of the Wilcoxon tests on the DNase data of the pairs
}
\keyword{datasets}
