% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dhsexp_test_data}
\alias{dhsexp_test_data}
\title{DHS expression test}
\format{
A data frame with three variables: \code{gene_id} Gene ID,
\code{symbol38} Enhancer ID, \code{wilcoxtest_dhs_exp} result of the test ).
}
\usage{
dhsexp_test_data
}
\description{
Results of the Wilcoxon test on the DNase data and RNA expression of the pairs
}
\keyword{datasets}
