% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distances.R
\name{distances}
\alias{distances}
\title{Compute genomic distances}
\usage{
distances(x)
}
\arguments{
\item{x}{A dataframe of either two columns (gene and enhancer pairs) or
one colum of genes}
}
\value{
A dataframe with the gene and enhancer pairs and the genomic distance
between them
}
\description{
Compute genomic distances
}
\examples{


}
