% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPrecomputedData.R
\name{downloadPrecomputedData}
\alias{downloadPrecomputedData}
\title{Download PrecomputedData.db}
\usage{
downloadPrecomputedData(method)
}
\arguments{
\item{method}{Method to be used for downloading files. Check download.file()
manuals to see the current available methods}
}
\value{
Integer code 0 if download was succesful and the file was saved in the
correct directory. In the case of failure errors will be raised.
}
\description{
Downloads the PrecomputedData.db database that is needed for the
computeGenericFeatures() function. This databased contains the values
of the Wilcoxon Rank tests and the CRUP correlations.
}
\examples{
# methods accepts "internal", "wininet" (Windows only) ,"wget", "curl" etc.
# Assuming the user has wget available
downloadPrecomputedData(method="wget")
}
