% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrePrediction.R
\name{centrePrediction}
\alias{centrePrediction}
\title{CENTRE predicton}
\usage{
centrePrediction(features_celltype, features_generic, model = NULL)
}
\arguments{
\item{features_celltype}{The cell type specific features returned by
\code{CENTRE::computeCellTypeFeatures()}}

\item{features_generic}{The generic features returned by the function
\code{CENTRE::computeGenericFeatures()}}

\item{model}{Path to the model the predictions will be computed on,
The default is the CENTRE model.}
}
\value{
Data.frame containing the enhancer gene pairs and the probability of
them interacting based on CENTRE model
}
\description{
Predicts if the enhancer gene pairs are interacting from the generic
features and cell type features.
}
\examples{
pairs <- data.frame(
    gene_id1 = c("ENSG00000105281"),
    enhancer_id = c("EH38E1958626")
)

generic_features <- CENTRE::computeGenericFeatures(pairs)

# Compute Cell-type features
eh <- ExperimentHub::ExperimentHub()

files <- c(
    system.file("extdata",
        "example/HeLa_H3K4me1.REF_chr19_reduced.bam",
        package = "CENTRE"
    ),
    system.file("extdata",
        "example/HeLa_H3K4me3.REF_chr19_reduced.bam",
        package = "CENTRE"
    ),
    system.file("extdata",
        "example/HeLa_H3K4me3.REF_chr19_reduced.bam",
        package = "CENTRE"
    )
)
inputs <- system.file("extdata",
    "example/HeLa_input.REF_chr19_reduced.bam",
    package = "CENTRE"
)
metaData <- data.frame(
    HM = c("H3K4me1", "H3K4me3", "H3K27ac"),
    condition = c(1, 1, 1), replicate = c(1, 1, 1),
    bamFile = files, inputFile = rep(inputs, 3)
)

tpmpath <- unname(eh[["EH9545"]])
tpmfile <- read.table(tpmpath,
    sep = "",
    stringsAsFactors = FALSE, header = TRUE
)
tpmfile <- tpmfile[grep("E", tpmfile$gene_id), ]
celltype_features <- CENTRE::computeCellTypeFeatures(metaData,
    replicate = 1,
    input.free = FALSE,
    cores = 1,
    sequencing = "single",
    tpmData = tpmfile,
    pairs = pairs
)
# Finally compute the predictions
predictions <- centrePrediction(celltype_features, generic_features)

}
