% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrePrediction.R
\name{centrePrediction}
\alias{centrePrediction}
\title{CENTRE predicton}
\usage{
centrePrediction(features_celltype, features_generic, model = NULL)
}
\arguments{
\item{features_celltype}{The cell type specific features returned by
\code{CENTRE::computeCellTypeFeatures()}}

\item{features_generic}{The generic features returned by the function
\code{CENTRE::computeGenericFeatures()}}

\item{model}{Path to the model the predictions will be computed on,
The default is the CENTRE model.}
}
\value{
Dataframe containing the enhancer gene pairs and the probability of
them interacting based on CENTRE model
}
\description{
CENTRE predicton
}
