% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrePrediction.R
\name{centrePrediction}
\alias{centrePrediction}
\title{CENTRE predicton}
\usage{
centrePrediction(features_celltype, features_generic, model = NULL)
}
\arguments{
\item{features_celltype}{The cell type specific features returned by
\code{CENTRE::computeCellTypeFeatures()}}

\item{features_generic}{The generic features returned by the function
\code{CENTRE::computeGenericFeatures()}}

\item{model}{Path to the model the predictions will be computed on,
The default is the CENTRE model.}
}
\value{
Dataframe containing the enhancer gene pairs and the probability of
them interacting based on CENTRE model
}
\description{
CENTRE predicton
}
\examples{
#Create gene enhancer pairs
genes <- as.data.frame(c("ENSG00000130203.10",
"ENSG00000171119.3"))
colnames(genes) <- c("gene_id") #It is important to name the column gene_id
pairs <- CENTRE::createPairs(genes)

#Compute generic features
colnames(pairs) <- c("gene_id", "enhancer_id")
generic_features <- CENTRE::computeGenericFeatures(pairs)

#Compute Cell-type features
files <- c(system.file("extdata/example","HeLa_H3K4me1.REF_chr19.bam", package = "CENTRE"),
system.file("extdata/example","HeLa_H3K4me3.REF_chr19.bam", package = "CENTRE"),
system.file("extdata/example","HeLa_H3K27ac.REF_chr19.bam", package = "CENTRE"))
# Control ChIP-seq experiment to go with the rest of ChIP-seqs
inputs <- system.file("extdata/example", "HeLa_input.REF_chr19.bam", package = "CENTRE")
metaData <- data.frame(HM = c("H3K4me1", "H3K4me3", "H3K27ac"),
                     condition = c(1, 1, 1), replicate = c(1, 1, 1),
                      bamFile = files, inputFile = rep(inputs, 3))
tpmfile <- read.table(system.file("extdata/example", "HeLa-S3.tsv", package = "CENTRE"),
                      sep = "", stringsAsFactors = FALSE, header = TRUE)
celltype_features <- CENTRE::computeCellTypeFeatures(metaData,
                                                    replicate = 1,
                                                    input.free = FALSE,
                                                    cores = 1,
                                                    sequencing = "single",
                                                    tpmfile = tpmfile,
                                                    pairs = pairs)
# Finally compute the predictions
predictions <- centrePrediction(celltype_features,
generic_features)
}
