% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCellTypeFeatures.R
\name{computeCellTypeFeatures}
\alias{computeCellTypeFeatures}
\title{Compute cell type specific features}
\usage{
computeCellTypeFeatures(
  metaData,
  replicate,
  input.free = FALSE,
  cores,
  sequencing = "single",
  tpmData,
  chr = NULL,
  pairs
)
}
\arguments{
\item{metaData}{Data.frame indicating the paths to the ChIP-seq experiments.
More information on the format here `crupR::normalize}

\item{replicate}{The number of replicates of the ChIP-seq experiments
that need to be normalized.}

\item{input.free}{Boolean value indicating whether a Control/Input ChIP-seq
experiment is provided to go with the Histone Modification ChIP-seq
experiments.
If the parameter is set to FALSE the normalization of ChIP-seq experiments
will be run in input.free mode.}

\item{cores}{Number of cores to compute the CRUP score features}

\item{sequencing}{Type of sequencing of the ChIP-seq experiments "paired" or
"single". The parameter takes single as default}

\item{tpmData}{Dataframe of two columns one with the RNA-seq TPM values,
one with the names of the genes given as ENSEMBLE ID's}

\item{chr}{NULL or a vector of chromosomes. Use only if the crupR
normalization should be done for certain chromosomes. If this parameter is
not used crupR normalization is done for all chromosomes.
Using only certain chromosomes for normalization might change results
and is not the intented used of crupR or CENTRE.}

\item{pairs}{The output of \code{CENTRE::createPairs()}.}
}
\value{
A table containing the following computed features :
\itemize{
\item CRUP enhancer score for enhancer region, promoter region and the region
between the enhancer and the promoter.
\item CRUP promoter score for enhancer region, promoter region and the region
between the enhancer and the promoter.
\item TPM values from the RNA-seq experiment given.
}
}
\description{
Computes the cell type specific features needed for the CENTRE classification
step.
}
\examples{
#for the sake of runtime we load pairs that
#were precomputed for the example. The user should run createPairs()
#to get them.
\dontrun{
pairs <- readRDS(file = system.file("extdata",
       "input_cellType_pairs.rds",
       package = "CENTRE"
   ))

# Compute Cell-type features
eh <- ExperimentHub::ExperimentHub()

files <- c(
    system.file("extdata",
        "example/HeLa_H3K4me1.REF_chr19_reduced.bam",
        package = "CENTRE"
    ),
    system.file("extdata",
        "example/HeLa_H3K4me3.REF_chr19_reduced.bam",
        package = "CENTRE"
    ),
    system.file("extdata",
        "example/HeLa_H3K4me3.REF_chr19_reduced.bam",
        package = "CENTRE"
    )
)
inputs <- system.file("extdata",
    "example/HeLa_input.REF_chr19_reduced.bam",
    package = "CENTRE"
)
metaData <- data.frame(
    HM = c("H3K4me1", "H3K4me3", "H3K27ac"),
    condition = c(1, 1, 1), replicate = c(1, 1, 1),
    bamFile = files, inputFile = rep(inputs, 3)
)

tpmpath <- unname(eh[["EH9545"]])
tpmfile <- read.table(tpmpath,
    sep = "", stringsAsFactors = FALSE,
    header = TRUE
)
tpmfile <- tpmfile[grep("E", tpmfile$gene_id), ]
celltype_features <- CENTRE::computeCellTypeFeatures(metaData,
    replicate = 1,
    input.free = FALSE,
    cores = 1,
    chr = "chr19",
    sequencing = "single",
    tpmData = tpmfile,
    pairs = pairs
)
}
}
