% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenericFeatures.R
\name{computeGenericFeatures}
\alias{computeGenericFeatures}
\title{Compute generic features}
\usage{
computeGenericFeatures(x)
}
\arguments{
\item{x}{Dataframe with the gene-enhancer pairs of interest}
}
\value{
A table containting the following computed features :
\itemize{
\item distance: Distance between gene and enhancer
\item combined_tests: Combined value of the Wilcoxon tests (CAGE, DNase
expression, CRUP expression and DNase DNase)
\item crup_COR: CRUP correlation scores
}
}
\description{
Computes the generic features needed by CENTRE to make predictions
}
\examples{
#Create candidate pairs
candidates <- read.table(system.file("extdata",
"exampleids.txt", package = "CENTRE"), header = T)
#Remember to give the columns the name "gene_id"
colnames(candidates) <- c("gene_id")
#Generate the candidate pairs
candidate_pairs <- createPairs(candidates)

#Or give your own
#Rememeber to name them correctly
colnames(candidate_pairs) <- c("gene_id", "enhancer_id")

generic_features <- computeGenericFeatures(candidate_pairs)

}
