% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenericFeatures.R
\name{computeGenericFeatures}
\alias{computeGenericFeatures}
\title{Compute generic features}
\usage{
computeGenericFeatures(pairs)
}
\arguments{
\item{pairs}{Dataframe with the gene-enhancer pairs of interest}
}
\value{
A table containting the following computed features :
\itemize{
\item distance: Distance between gene and enhancer
\item combined_tests: Combined value of the Wilcoxon tests (CAGE, DNase
expression, CRUP expression and DNase DNase)
\item crup_COR: CRUP correlation scores
}
}
\description{
Computes the generic features needed by CENTRE to make predictions
}
\examples{
#Create gene enhancer pairs
genes <- as.data.frame(c("ENSG00000130203.10",
"ENSG00000171119.3"))
colnames(genes) <- c("gene_id") #It is important to name the column gene_id
pairs <- CENTRE::createPairs(genes)

#Compute generic features
colnames(pairs) <- c("gene_id", "enhancer_id")
generic_features <- CENTRE::computeGenericFeatures(pairs)
}
