% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGenericFeatures.R
\name{computeGenericFeatures}
\alias{computeGenericFeatures}
\title{Compute generic features}
\usage{
computeGenericFeatures(x)
}
\arguments{
\item{file}{Path to a file with either genes and enhancer pairs.}

\item{metaData}{Dataframe indicating the paths to the ChIP-seq experiments.
More information on the format here \code{crupR::normalize()}}
}
\value{
A table containting the following computed features :
\itemize{
\item distance: Distance between gene and enhancer
\item combined_tests: Combined value of the Wilcoxon tests (CAGE, DNase
expression, CRUP expression and DNase DNase)
\item crup_COR: CRUP correlation scores
}
}
\description{
Computes the generic features needed by CENTRE to make predictions
}
