% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPairs.R
\name{createPairs}
\alias{createPairs}
\title{Create Pairs}
\usage{
createPairs(gene)
}
\arguments{
\item{gene}{One column dataframe with gene ENSEMBL id's}
}
\value{
dataframe with two columns the ENSEMBL id's without their version and
the enhancer id's
}
\description{
Creates all of the possible gene enhancer pairs at 500kb distance
from the given genes
transcription start sites.
}
\examples{
#Start by providing genes with their ENSEMBL id
candidates <- read.table(system.file("extdata",
"exampleids.txt", package = "CENTRE"), header = T)
#Remember to give the columns the name "gene_id"
colnames(candidates) <- c("gene_id")
#Generate the candidate pairs
candidate_pairs <- createPairs(candidates)
}
