% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPairs.R
\name{createPairs}
\alias{createPairs}
\title{Create Pairs}
\usage{
createPairs(gene)
}
\arguments{
\item{gene}{One column dataframe with gene ENSEMBL id's}
}
\value{
dataframe with two columns the ENSEMBL id's without their version and
the enhancer id's
}
\description{
Creates all of the possible gene enhancer pairs at 500kb distance
from the given genes
transcription start sites.
}
\examples{
#Create gene enhancer pairs
genes <- as.data.frame(c("ENSG00000130203.10",
"ENSG00000171119.3"))
colnames(genes) <- c("gene_id") #It is important to name the column gene_id
pairs <- CENTRE::createPairs(genes)
}
