% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crup_cor}
\alias{crup_cor}
\title{CRUP correlations}
\format{
A data frame with three variables: \code{gene_id} Gene ID,
\code{symbol38} Enhancer ID, \code{cor_CRUP} value of the correlation ).
}
\usage{
crup_cor
}
\description{
Correlation between the CRUP scores of enhancer and genes  in the pairs
}
\keyword{datasets}
