% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crupexp_test_data}
\alias{crupexp_test_data}
\title{CRUP expression test}
\format{
A data frame with five variables: \code{year}, \code{sex},
\code{name}, \code{n} and \code{prop} (\code{n} divided by total number
of applicants in that year, which means proportions are of people of
that sex with that name born in that year).
}
\usage{
crupexp_test_data
}
\description{
Correlation between the CRUP scores of enhancer and genes  in the pairs
}
\keyword{datasets}
